requirements.txt
Flask==2.3.3
pandas==2.2.1
openpyxl==3.1.2
gunicorn>=23.0.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
requests
python-dotenv
gspread==5.10.0
google-auth==2.23.4
cloudinary==1.36.0
